django_project/requirements.txt
django
geopandas
djangorestframework
